+++ .vscode/settings.json
@@ -37,6 +37,7 @@
         "gainsboro",
         "GANTT",
         "Gggpound",
+        "Govett",
         "Gridlines",
         "Hanalei",
         "hdmx",
@@ -69,9 +70,9 @@
         "mengshukeji",
         "Neue",
         "numfmt",
+        "OOXML",
         "opencollective",
         "opentype",
-        "OOXML",
         "Overlines",
         "Pacifico",
         "Plass",
